Added courses, teachers and students
