+++ src/components/courses/CoursesModal.js
@@ -3,7 +3,19 @@ import { styled } from '@mui/material/styles';
 import Dialog from '@mui/material/Dialog';
 import DialogContent from '@mui/material/DialogContent';
 import { Box } from '@mui/material';
-import ManageCourseForm from './ManageCourseForm';
+import ManageCourseForm from '../courses/ManageCourseForm';
+import { makeStyles } from '@material-ui/core'
+
+const useStyles = makeStyles(() => ({
+  button: {
+    width: '110px',
+    height: '40px',
+    color: 'black',
+  },
+  buttonName: {
+    color: 'black',
+  }
+}))
 
 const BootstrapDialog = styled(Dialog)(({ theme }) => ({
   '& .MuiDialogContent-root': {
@@ -15,6 +27,7 @@ const BootstrapDialog = styled(Dialog)(({ theme }) => ({
 }));
 
 export default function CustomizedDialogs({open, setOpen, courses, loadData, editForm, handleCloseModal, setEditForm}) {
+  const classes = useStyles()
   return (
     <Box>
       <Button  
@@ -22,8 +35,10 @@ export default function CustomizedDialogs({open, setOpen, courses, loadData, edi
           setOpen(true)
           setEditForm({})}
         }
+        variant="outlined"
+        className={classes.button}
       >
-        Add New
+        <p className={classes.buttonName}>Add New</p>
       </Button>
       <BootstrapDialog
         onClose={() => setOpen(false)}
