+++ src/components/students/Students.js
@@ -1,9 +1,9 @@
-import TeachersModal from './TeachersModal'
+import StudentsModal from '../students/StudentsModal'
 import { Box } from '@mui/material'
 import { makeStyles } from '@material-ui/core'
 import { useState, useEffect } from 'react'
-import TableTeachers from './TableTeachers'
-import { getTeachers } from '../Services'
+import TableStudents from '../students/TableStudents'
+import { getStudents } from '../../services/StudentsService'
 
 const useStyles = makeStyles((theme) => ({
   table: {
@@ -16,7 +16,7 @@ const useStyles = makeStyles((theme) => ({
     [theme.breakpoints.down('lg')]: {
       paddingLeft: '200px',
       paddingTop: theme.spacing(7),
-      width: '1460px',
+      width: '1450px',
     },
     [theme.breakpoints.down('md')]: {
       paddingLeft: '200px',
@@ -34,12 +34,47 @@ const useStyles = makeStyles((theme) => ({
       minWidth: '260px'
     }
   },
-  titlePage: {
-    textShadow: '0 0 3px blue',
-    color: 'white',
-    width: '94%',
+  tableHide: {
+    [theme.breakpoints.down('xl')]: {
+      paddingLeft: '10px',
+      paddingTop: theme.spacing(7),
+      minWidth: '260px',
+      paddingBottom: '80px'
+    },
+    [theme.breakpoints.down('lg')]: {
+      paddingLeft: '10px',
+      paddingTop: theme.spacing(7),
+      width: '1640px',
+    },
+    [theme.breakpoints.down('md')]: {
+      paddingLeft: '10px',
+      paddingTop: theme.spacing(7),
+      width: '1230px',
+    },
+    [theme.breakpoints.down('sm')]: {
+      paddingLeft: '10px',
+      paddingTop: theme.spacing(7),
+      width: '900px',
+    },
     [theme.breakpoints.down('xs')]: {
-      width: '69%',
+      paddingLeft: '2px',
+      paddingTop: theme.spacing(6),
+      minWidth: '260px'
+    }
+  },
+  titlePage: {
+    color: 'black',
+    fontWeight: '500',
+    fontFamily: 'Roboto',
+    width: '92.5%',
+    [theme.breakpoints.down('xs')]: {
+      width: '59%',
+    }
+  },
+  titlePageHide: {
+    width: '93.4%',
+    [theme.breakpoints.down('xs')]: {
+      width: '70%',
     }
   },
   header: {
@@ -47,35 +82,36 @@ const useStyles = makeStyles((theme) => ({
   },
   modal: {
     paddingTop: theme.spacing(2.2),
+    width: '100px'
   }
 }))
 
-const Teachers = () => {
+const Students = ({show}) => {
   const [open, setOpen] = useState(false)
-  const [teachers, setTeachers] = useState([])
+  const [students, setStudents] = useState([])
   useEffect(() => {
-    loadTeachers()
+    loadStudents()
   }, [])
-  const loadTeachers = () => getTeachers().then((teacher) => setTeachers(teacher))
+  const loadStudents = () => getStudents().then((student) => setStudents(student))
   const [editForm, setEditForm] = useState({});
   const handleCloseModal = () => {
     setOpen(false)
   }
   const loadData = () => {
     handleCloseModal()
-    loadTeachers()
+    loadStudents()
   }
-  const openEditModal = (teacher) => {
-    setEditForm(teacher)
+  const openEditModal = (student) => {
+    setEditForm(student)
     setOpen(true)
   }
   const classes = useStyles()
   return (
-    <Box className={classes.table}>
+    <Box className={show ? classes.table : classes.tableHide}>
       <Box className={classes.header}>
-        <h2 className={classes.titlePage}>Teachers</h2>
+        <h2 className={show ? classes.titlePage : classes.titlePageHide}>Students</h2>
         <Box className={classes.modal}>
-          <TeachersModal 
+          <StudentsModal 
             open = {open}
             setOpen = {setOpen}
             loadData = {loadData} 
@@ -85,8 +121,8 @@ const Teachers = () => {
           />  
         </Box>
       </Box>
-      <TableTeachers 
-        teachers = {teachers}
+      <TableStudents 
+        students = {students}
         loadData = {loadData} 
         openEditModal = {openEditModal}
       />
@@ -94,4 +130,4 @@ const Teachers = () => {
   )
 }
 
-export default Teachers
+export default Students
