+++ src/components/teachers/Teachers.js
@@ -1,9 +1,9 @@
-import StudentsModal from './StudentsModal'
+import TeachersModal from '../teachers/TeachersModal'
 import { Box } from '@mui/material'
 import { makeStyles } from '@material-ui/core'
 import { useState, useEffect } from 'react'
-import TableStudents from './TableStudents'
-import { getStudents } from '../Services'
+import TableTeachers from '../teachers/TableTeachers'
+import { getTeachers } from '../../services/TeachersService'
 
 const useStyles = makeStyles((theme) => ({
   table: {
@@ -16,7 +16,7 @@ const useStyles = makeStyles((theme) => ({
     [theme.breakpoints.down('lg')]: {
       paddingLeft: '200px',
       paddingTop: theme.spacing(7),
-      width: '1460px',
+      width: '1450px',
     },
     [theme.breakpoints.down('md')]: {
       paddingLeft: '200px',
@@ -34,12 +34,47 @@ const useStyles = makeStyles((theme) => ({
       minWidth: '260px'
     }
   },
-  titlePage: {
-    textShadow: '0 0 3px blue',
-    color: 'white',
-    width: '94%',
+  tableHide: {
+    [theme.breakpoints.down('xl')]: {
+      paddingLeft: '10px',
+      paddingTop: theme.spacing(7),
+      minWidth: '260px',
+      paddingBottom: '80px'
+    },
+    [theme.breakpoints.down('lg')]: {
+      paddingLeft: '10px',
+      paddingTop: theme.spacing(7),
+      width: '1640px',
+    },
+    [theme.breakpoints.down('md')]: {
+      paddingLeft: '10px',
+      paddingTop: theme.spacing(7),
+      width: '1230px',
+    },
+    [theme.breakpoints.down('sm')]: {
+      paddingLeft: '10px',
+      paddingTop: theme.spacing(7),
+      width: '900px',
+    },
     [theme.breakpoints.down('xs')]: {
-      width: '69%',
+      paddingLeft: '2px',
+      paddingTop: theme.spacing(6),
+      minWidth: '260px'
+    }
+  },
+  titlePage: {
+    color: 'black',
+    fontWeight: '500',
+    fontFamily: 'Roboto',
+    width: '92.5%',
+    [theme.breakpoints.down('xs')]: {
+      width: '59%',
+    }
+  },
+  titlePageHide: {
+    width: '93.4%',
+    [theme.breakpoints.down('xs')]: {
+      width: '70%',
     }
   },
   header: {
@@ -47,35 +82,36 @@ const useStyles = makeStyles((theme) => ({
   },
   modal: {
     paddingTop: theme.spacing(2.2),
+    width: '100px'
   }
 }))
 
-const Students = () => {
+const Teachers = ({show}) => {
   const [open, setOpen] = useState(false)
-  const [students, setStudents] = useState([])
+  const [teachers, setTeachers] = useState([])
   useEffect(() => {
-    loadStudents()
+    loadTeachers()
   }, [])
-  const loadStudents = () => getStudents().then((student) => setStudents(student))
+  const loadTeachers = () => getTeachers().then((teacher) => setTeachers(teacher))
   const [editForm, setEditForm] = useState({});
   const handleCloseModal = () => {
     setOpen(false)
   }
   const loadData = () => {
     handleCloseModal()
-    loadStudents()
+    loadTeachers()
   }
-  const openEditModal = (student) => {
-    setEditForm(student)
+  const openEditModal = (teacher) => {
+    setEditForm(teacher)
     setOpen(true)
   }
   const classes = useStyles()
   return (
-    <Box className={classes.table}>
+    <Box className={show ? classes.table : classes.tableHide}>
       <Box className={classes.header}>
-        <h2 className={classes.titlePage}>Students</h2>
+        <h2 className={show ? classes.titlePage : classes.titlePageHide}>Teachers</h2>
         <Box className={classes.modal}>
-          <StudentsModal 
+          <TeachersModal 
             open = {open}
             setOpen = {setOpen}
             loadData = {loadData} 
@@ -85,8 +121,8 @@ const Students = () => {
           />  
         </Box>
       </Box>
-      <TableStudents 
-        students = {students}
+      <TableTeachers 
+        teachers = {teachers}
         loadData = {loadData} 
         openEditModal = {openEditModal}
       />
@@ -94,4 +130,4 @@ const Students = () => {
   )
 }
 
-export default Students
+export default Teachers
